refactor: extract classifyContest to classification.ts
